Fixed #BA1325 caused by an IntegerOverflow. Due to this, changed the return type of getPrice() to long. https://beyond-aion.com/forum/thread/1325-ruhnstamm-ausr%C3%BCstung-verkaufen/?postID=4446#post4446

+++ game-server/src/com/aionemu/gameserver/model/templates/item/ItemTemplate.java
@@ -239,7 +239,7 @@ public class ItemTemplate extends VisibleObjectTemplate {
 		return itemGroup.getEquipType();
 	}
 
-	public int getPrice() {
+	public long getPrice() {
 		return price;
 	}
